[Cython] exctract import data methods for cythonization
