+++ octobot_backtesting/importers/__init__.pxd
@@ -25,9 +25,21 @@ from octobot_backtesting.importers cimport exchanges
 
 from octobot_backtesting.importers.exchanges cimport (
     ExchangeDataImporter,
+    get_operations_from_timestamps,
+    import_ohlcvs,
+    import_tickers,
+    import_order_books,
+    import_recent_trades,
+    import_klines,
 )
 
 __all__ = [
     "DataImporter",
     "ExchangeDataImporter",
+    "get_operations_from_timestamps",
+    "import_ohlcvs",
+    "import_tickers",
+    "import_order_books",
+    "import_recent_trades",
+    "import_klines",
 ]
