+++ octobot_backtesting/importers/exchanges/exchange_importer.py
@@ -13,8 +13,6 @@
 #
 #  You should have received a copy of the GNU Lesser General Public
 #  License along with this library.
-import json
-
 import octobot_commons.constants as common_constants
 import octobot_commons.enums as common_enums
 
@@ -102,132 +100,86 @@ class ExchangeDataImporter(importers.DataImporter):
                                      if await self.database.check_table_exists(table)
                                      and await self.database.check_table_not_empty(table)]
 
-    @staticmethod
-    def get_operations_from_timestamps(superior_timestamp, inferior_timestamp):
-        operations: list = []
-        timestamps: list = []
-        if superior_timestamp != -1:
-            timestamps.append(str(superior_timestamp))
-            operations.append(enums.DataBaseOperations.INF_EQUALS.value)
-        if inferior_timestamp != -1:
-            timestamps.append(str(inferior_timestamp))
-            operations.append(enums.DataBaseOperations.SUP_EQUALS.value)
-
-        return timestamps, operations
-
-    @staticmethod
-    def import_ohlcvs(ohlcvs):
-        for i, val in enumerate(ohlcvs):
-            ohlcvs[i] = list(val)
-            # ohlcvs[i][-2] = TimeFrames(ohlcvs[i][-2])
-            ohlcvs[i][-1] = json.loads(ohlcvs[i][-1])
-        return ohlcvs
-
     async def get_ohlcv(self, exchange_name=None, symbol=None,
                         time_frame=common_enums.TimeFrames.ONE_HOUR,
                         limit=data.DataBase.DEFAULT_SIZE):
-        return ExchangeDataImporter.import_ohlcvs(await self.database.select(enums.ExchangeDataTables.OHLCV, size=limit,
-                                                                             exchange_name=exchange_name, symbol=symbol,
-                                                                             time_frame=time_frame.value))
+        return importers.import_ohlcvs(await self.database.select(enums.ExchangeDataTables.OHLCV, size=limit,
+                                                                  exchange_name=exchange_name, symbol=symbol,
+                                                                  time_frame=time_frame.value))
 
     async def get_ohlcv_from_timestamps(self, exchange_name=None, symbol=None,
                                         time_frame=common_enums.TimeFrames.ONE_HOUR,
                                         limit=data.DataBase.DEFAULT_SIZE,
                                         inferior_timestamp=-1, superior_timestamp=-1) -> list:
-        timestamps, operations = ExchangeDataImporter.get_operations_from_timestamps(superior_timestamp,
-                                                                                     inferior_timestamp)
-        return ExchangeDataImporter.import_ohlcvs(await self.database.select_from_timestamp(
+        timestamps, operations = importers.get_operations_from_timestamps(superior_timestamp,
+                                                                          inferior_timestamp)
+        return importers.import_ohlcvs(await self.database.select_from_timestamp(
             enums.ExchangeDataTables.OHLCV, size=limit,
             exchange_name=exchange_name, symbol=symbol,
             time_frame=time_frame.value,
             timestamps=timestamps,
             operations=operations))
 
-    @staticmethod
-    def import_tickers(tickers):
-        for ticker in tickers:
-            ticker[-1] = json.loads(ticker[-1])
-        return tickers
-
     async def get_ticker(self, exchange_name=None, symbol=None, limit=data.DataBase.DEFAULT_SIZE):
-        return ExchangeDataImporter.import_tickers(
+        return importers.import_tickers(
             await self.database.select(enums.ExchangeDataTables.TICKER, size=limit,
                                        exchange_name=exchange_name, symbol=symbol))
 
     async def get_ticker_from_timestamps(self, exchange_name=None, symbol=None, limit=data.DataBase.DEFAULT_SIZE,
                                          inferior_timestamp=-1, superior_timestamp=-1):
-        timestamps, operations = ExchangeDataImporter.get_operations_from_timestamps(superior_timestamp,
-                                                                                     inferior_timestamp)
-        return ExchangeDataImporter.import_tickers(
+        timestamps, operations = importers.get_operations_from_timestamps(superior_timestamp,
+                                                                          inferior_timestamp)
+        return importers.import_tickers(
             await self.database.select_from_timestamp(enums.ExchangeDataTables.TICKER, size=limit,
                                                       exchange_name=exchange_name, symbol=symbol,
                                                       timestamps=timestamps,
                                                       operations=operations))
 
-    @staticmethod
-    def import_order_books(order_books):
-        for order_book in order_books:
-            order_book[-1] = json.loads(order_book[-1])
-            order_book[-2] = json.loads(order_book[-2])
-        return order_books
-
     async def get_order_book(self, exchange_name=None, symbol=None, limit=data.DataBase.DEFAULT_SIZE):
-        return ExchangeDataImporter.import_order_books(
+        return importers.import_order_books(
             await self.database.select(enums.ExchangeDataTables.ORDER_BOOK, size=limit,
                                        exchange_name=exchange_name, symbol=symbol))
 
     async def get_order_book_from_timestamps(self, exchange_name=None, symbol=None, limit=data.DataBase.DEFAULT_SIZE,
                                              inferior_timestamp=-1, superior_timestamp=-1):
-        timestamps, operations = ExchangeDataImporter.get_operations_from_timestamps(superior_timestamp,
-                                                                                     inferior_timestamp)
-        return ExchangeDataImporter.import_order_books(
+        timestamps, operations = importers.get_operations_from_timestamps(superior_timestamp,
+                                                                          inferior_timestamp)
+        return importers.import_order_books(
             await self.database.select_from_timestamp(enums.ExchangeDataTables.ORDER_BOOK, size=limit,
                                                       exchange_name=exchange_name, symbol=symbol,
                                                       timestamps=timestamps, operations=operations))
 
-    @staticmethod
-    def import_recent_trades(recent_trades):
-        for recent_trade in recent_trades:
-            recent_trade[-1] = json.loads(recent_trade[-1])
-        return recent_trades
-
     async def get_recent_trades(self, exchange_name=None, symbol=None, limit=data.DataBase.DEFAULT_SIZE):
-        return ExchangeDataImporter.import_recent_trades(
+        return importers.import_recent_trades(
             await self.database.select(enums.ExchangeDataTables.RECENT_TRADES, size=limit,
                                        exchange_name=exchange_name, symbol=symbol))
 
     async def get_recent_trades_from_timestamps(self, exchange_name=None, symbol=None, limit=data.DataBase.DEFAULT_SIZE,
                                                 inferior_timestamp=-1, superior_timestamp=-1):
-        timestamps, operations = ExchangeDataImporter.get_operations_from_timestamps(superior_timestamp,
-                                                                                     inferior_timestamp)
-        return ExchangeDataImporter.import_recent_trades(await
-                                                         self.database.select_from_timestamp(
-                                                             enums.ExchangeDataTables.RECENT_TRADES,
-                                                             size=limit,
-                                                             exchange_name=exchange_name,
-                                                             symbol=symbol,
-                                                             timestamps=timestamps,
-                                                             operations=operations))
-
-    @staticmethod
-    def import_klines(klines):
-        for kline in klines:
-            kline[-1] = json.loads(kline[-1])
-        return klines
+        timestamps, operations = importers.get_operations_from_timestamps(superior_timestamp,
+                                                                          inferior_timestamp)
+        return importers.import_recent_trades(await
+                                              self.database.select_from_timestamp(
+                                                  enums.ExchangeDataTables.RECENT_TRADES,
+                                                  size=limit,
+                                                  exchange_name=exchange_name,
+                                                  symbol=symbol,
+                                                  timestamps=timestamps,
+                                                  operations=operations))
 
     async def get_kline(self, exchange_name=None, symbol=None,
                         time_frame=common_enums.TimeFrames.ONE_HOUR, limit=data.DataBase.DEFAULT_SIZE):
-        return ExchangeDataImporter.import_klines(await self.database.select(enums.ExchangeDataTables.KLINE, size=limit,
-                                                                             exchange_name=exchange_name, symbol=symbol,
-                                                                             time_frame=time_frame.value))
+        return importers.import_klines(await self.database.select(enums.ExchangeDataTables.KLINE, size=limit,
+                                                                  exchange_name=exchange_name, symbol=symbol,
+                                                                  time_frame=time_frame.value))
 
     async def get_kline_from_timestamps(self, exchange_name=None, symbol=None,
                                         time_frame=common_enums.TimeFrames.ONE_HOUR,
                                         limit=data.DataBase.DEFAULT_SIZE,
                                         inferior_timestamp=-1, superior_timestamp=-1):
-        timestamps, operations = ExchangeDataImporter.get_operations_from_timestamps(superior_timestamp,
-                                                                                     inferior_timestamp)
-        return ExchangeDataImporter.import_klines(
+        timestamps, operations = importers.get_operations_from_timestamps(superior_timestamp,
+                                                                          inferior_timestamp)
+        return importers.import_klines(
             await self.database.select_from_timestamp(enums.ExchangeDataTables.KLINE, size=limit,
                                                       exchange_name=exchange_name, symbol=symbol,
                                                       time_frame=time_frame.value,
